NXP-9610: setup integration tests with marketplace package

+++ tools/integration_tests_setup.py
@@ -1,19 +1,30 @@
 r"""Setup the windows machine for Continuous Integration on the MSI package
 
-This will:
+Steps executed by this script:
 
     - download the latest nightly build of Nuxeo from Jenkins
-    - deploy the server side package of nuxeo-drive (TODO)
+    - deploy the marketplace package of nuxeo-drive
     - start the server
     - setup test environment variables
+
+Under Windows:
+
     - download http://code.google.com/p/lessmsi if not already there
     - extract the MSI archive in a local, non system folder with lessmsi
     - launch the ``ndrive test`` command to run the integration tests
+
+Under POSIX:
+
+    - build the nuxeo-drive-client from the source using system python
+    - run the integration tests with nosetests
+
+Both:
+
     - stop the server and collect the logs
 
 Get the help on running this with::
 
-    C:\Python27\python tools\windows\integration_tests_setup.py --help
+    python integration_tests_setup.py --help
 
 """
 
@@ -28,7 +39,10 @@ import atexit
 import time
 
 
-DEFAULT_MSI = "dist"
+DEFAULT_MARKETPLACE = os.path.join(
+    "packaging", "nuxeo-drive-marketplace", "target")
+MARKET_PLACE_PREFIX = "nuxeo-drive-marketplace"
+DEFAULT_MSI = os.path.join(r"nuxeo-drive-client\dist")
 DEFAULT_NUXEO_ARCHIVE_URL=("http://qa.nuxeo.org/jenkins/job/IT-nuxeo-master-build/"
                            "lastSuccessfulBuild/artifact/archives/")
 DEFAULT_LESSMSI_URL="http://lessmsi.googlecode.com/files/lessmsi-v1.0.8.zip"
@@ -85,24 +99,39 @@ def unzip(filename, target=None):
             f.write(zf.read(info.filename))
 
 
+def find_latest(folder, prefix=None, suffix=None):
+    files = os.listdir(folder)
+    if prefix is not None:
+        files = [f for f in files if f.startswith(prefix)]
+    if suffix is not None:
+        files = [f for f in files if f.endswith(suffix)]
+    if not files:
+        raise RuntimeError(('Could not find file with prefix "%s"'
+                            'and suffix "%s" in "%s"')
+                           % (prefix, suffix, msi_filename))
+    files.sort()
+    return os.path.join(folder, files[-1])
+
+
 def setup_nuxeo(nuxeo_archive_url):
     try:
         java_home = os.environ['JAVA_HOME']
     except KeyError:
         raise RuntimeError('The JAVA_HOME env variable is required.')
-    java_bin = os.path.join(java_home, 'bin')
-    path = os.environ['PATH']
-    if not java_bin in path:
-        os.environ['PATH'] = path + ";" + java_bin
+    if sys.platform == 'win32':
+        java_bin = os.path.join(java_home, 'bin')
+        path = os.environ['PATH']
+        if not java_bin in path:
+            os.environ['PATH'] = path + ";" + java_bin
 
-    print "Kill previous soffice if any to unlock old files"
-    execute('taskkill /f /fi "imagename eq soffice.*"')
+        print "Kill previous soffice if any to unlock old files"
+        execute('taskkill /f /fi "imagename eq soffice.*"')
 
-    print "Kill any potential rogue instance of ndrive.exe"
-    execute('taskkill /f /fi "imagename eq ndrive.exe"')
+        print "Kill any potential rogue instance of ndrive.exe"
+        execute('taskkill /f /fi "imagename eq ndrive.exe"')
 
-    print "Waiting for any killed process to actually stop"
-    time.sleep(1.0)
+        print "Waiting for any killed process to actually stop"
+        time.sleep(1.0)
 
     print "Finding latest nuxeo ZIP archive at: " + nuxeo_archive_url
     index_html = urllib2.urlopen(nuxeo_archive_url).read()
@@ -122,6 +151,8 @@ def setup_nuxeo(nuxeo_archive_url):
     if os.path.exists(NUXEO_FOLDER):
         # stop any previous server process that could have been left running
         # if jenkins kills this script
+        if sys.platform != 'win32':
+            execute("chmod +x " + nuxeoctl)
         execute(nuxeoctl + " --gui false stop")
         shutil.rmtree(NUXEO_FOLDER)
 
@@ -129,8 +160,15 @@ def setup_nuxeo(nuxeo_archive_url):
     with open(os.path.join(NUXEO_FOLDER, 'bin', 'nuxeo.conf'), 'wb') as f:
         f.write("\nnuxeo.wizard.done=true\n")
 
+    if sys.platform != 'win32':
+        execute("chmod +x " + nuxeoctl)
+
+    print "Installing the nuxeo drive marketplace package"
+    package = find_latest(DEFAULT_MARKETPLACE, prefix=MARKET_PLACE_PREFIX,
+                          suffix=".zip")
+    execute(nuxeoctl + " mp-install --accept=true --nodeps " + package)
+
     print "Starting the Nuxeo server"
-    nuxeoctl = os.path.join(NUXEO_FOLDER, 'bin', 'nuxeoctl')
     execute(nuxeoctl + " --gui false start")
 
     # Register a callback to stop the nuxeo server
@@ -139,15 +177,11 @@ def setup_nuxeo(nuxeo_archive_url):
 
 def extract_msi(lessmsi_url, msi_filename):
     if os.path.isdir(msi_filename):
-        files = os.listdir(msi_filename)
-        files = [f for f in files if f.endswith('.msi')]
-        if not files:
-            raise RuntimeError('Could not file msi file in ' + msi_filename)
-        files.sort()
-        msi_filename = os.path.join(msi_filename, files[-1])
-    filename = os.path.basename(lessmsi_url)
+        msi_filename = find_latest(msi_filename, suffix='.msi')
+
     if not os.path.exists(LESSMSI_FOLDER):
         download(lessmsi_url, filename)
+        filename = os.path.basename(lessmsi_url)
         unzip(filename, target=LESSMSI_FOLDER)
 
     print "Extracting the MSI"
@@ -157,16 +191,27 @@ def extract_msi(lessmsi_url, msi_filename):
     execute("%s /x %s %s" % (lessmsi, msi_filename, EXTRACTED_MSI_FOLDER))
 
 
-def run_tests():
-    ndrive = os.path.join(EXTRACTED_MSI_FOLDER, 'SourceDir', 'ndrive.exe')
+def set_environment():
     os.environ['NXDRIVE_TEST_NUXEO_URL'] = "http://localhost:8080/nuxeo"
     os.environ['NXDRIVE_TEST_USER'] = "Administrator"
     os.environ['NXDRIVE_TEST_PASSWORD'] = "Administrator"
+
+
+def run_tests_from_msi():
+    ndrive = os.path.join(EXTRACTED_MSI_FOLDER, 'SourceDir', 'ndrive.exe')
     execute(ndrive + " test")
+
+def run_tests_from_source():
+    execute("cd nuxeo-drive-client && nosetests")
 
 
 if __name__ == "__main__":
     options = parse_args()
     setup_nuxeo(options.nuxeo_archive_url)
-    extract_msi(options.lessmsi_url, options.msi)
-    run_tests()
+    set_environment()
+    if sys.platform == 'win32':
+        extract_msi(options.lessmsi_url, options.msi)
+        run_tests_from_msi()
+    else:
+        run_tests_from_source()
+
